requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
pytest==8.3.4
pytest-cov==6.0.0
httpx==0.28.1
apsw==3.49.1.0
apswutils==0.0.2
bcrypt==4.0.1
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastcore==1.8.1
fastlite==0.1.3
fastui==0.7.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==25.0
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-fasthtml==0.12.14
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uv==0.6.14
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
limits==5.6.0
slowapi==0.1.9
wrapt==1.17.3
Pillow==11.0.0
smbprotocol==1.14.0
